search_source/container_extracter/test/app/requirements.txt
##############################################################################
# ─────────────────── 서비스 코드에서 실제로 “사용”하는 12개 ───────────────────
##############################################################################
Flask==1.1.2                  # [CVE-2018-1000656, CVE-2019-1010083] 디버그 RCE 등
pyyaml==5.1                   # [CVE-2020-14343] yaml.load() RCE PoC

MarkupSafe==1.1.1             # [CVE-2021-27291] escape() HTML 인젝션
Jinja2==2.10.1                # [CVE-2019-10906] |attr() XSS
urllib3==1.23.0               # [CVE-2019-11324] Redirect SSRF (requests 호환)
Werkzeug==0.15.5              # [CVE-2019-0201] Debug 핀 RCE
cryptography==2.8             # (업데이트) paramiko & pyOpenSSL 호환

python-multipart==0.0.5       # CVE 없음 (경계 처리 이슈 연구 보고)
itsdangerous==1.1.0           # CVE 없음 (비밀키 유출 시 세션 위조)
gunicorn==19.9.0              # CVE 없음 (구성 실수 시 헤더 주입)
simplejson==3.17.0            # CVE 없음 (eval 유사 함수 주의)
requests==2.19.1              # [CVE-2018-18074] 인증 우회 리다이렉트

##############################################################################
# ─── “코드에서 호출하지 않지만 설치”할 라이브러리 29개 (취약+안전 혼합) ───
##############################################################################
# ▸ 비교적 “취약”으로 알려진 또는 구버전
Django==1.11.20               # [CVE-2019-6975] ORM SQLi
idna==2.6                     # [CVE-2017-14062] 오버플로
SQLAlchemy==1.3.23            # [CVE-2020-15190] SQL 삽입
markdown==2.6.9               # [CVE-2018-20164] XSS
paramiko==2.1.6               # [CVE-2018-7750] DH 취약
docutils==0.15.2              # [CVE-2021-32728] XSS, botocore 호환(<0.16)
bleach==3.1.2                 # [CVE-2021-23980] 필터 우회
PyJWT==1.7.1                  # [CVE-2020-26160] None 알고리즘 우회
httplib2==0.18.0              # [CVE-2020-11078] 헤더 인젝션
celery==4.4.7                 # [CVE-2021-23727] 경로 탐색
feedparser==6.0.10            # 최신 Python 호환
xlrd==1.2.0                   # [CVE-2020-24584] Formula 인젝션
websocket-client==0.54.0      # [CVE-2020-27830] 인증서 우회
numpy==1.16.0                 # [CVE-2019-6446] 버퍼 오버플로

# ▸ “현재까지 CVE 보고 없거나 위험성 낮은” 안전 계열
chardet==3.0.4                # 문자 인코딩 탐지
pygments==2.15.1              # 코드 하이라이터 (신규 안전 버전)
typing-extensions==3.7.4      # 타입 지원 백포트
python-dateutil==2.8.1        # 날짜 처리
certifi==2019.11.28           # 루트 CA 번들
boto3==1.12.0                 # AWS SDK
pyOpenSSL==19.1.0             # SSL 래퍼 (cryptography>=2.8 필요)
six==1.11.0                   # 호환성 유틸
cachetools==4.2.4             # 캐시 헬퍼
pytz==2021.1                  # 타임존 DB
decorator==4.4.2              # 함수 데코레이터
rich==10.9.0                  # 컬러 터미널 출력
click==8.1.8                  # CLI 유틸
cryptography-vectors==2.3     # 테스트 벡터 패키지
